Changed Brightcove method calls back to string constants (rather than using an enum)
Changed the accessility of the set accessor for each model item that has an Id property to public.
This allows API consumers who want to instantiate a new item with a known ID for update to do so.

Changed the serialization logic for the playlist objects to work around Brightcove errors.

Changed the default TagInclusion for playlist to None (which defaults to Or when submitted to Brightcove)

Added integration tests